fix(tests): address PR #2714 review findings 1, 3, 4, 5
- Finding 1 (IPC probe): supports_ipc_mempool inspected the raw
  CUresult and special-cased CUDA_ERROR_NOT_SUPPORTED. Ralf says
  unsupported handle types are represented by a successful bitmask without the
  POSIX-FD bit, so the special case is unnecessary. Remove it; let
  handle_return raise on any non-success and rely on the bitmask check.
  Update AGENTS.md example to match.

- Finding 3 (headless EGL): is_gl_context_unavailable rejected MissingFunctionException
  (pyglet raises it when libEGL exists but eglQueryDevicesEXT /
  eglGetPlatformDisplayEXT entry points do not). Add it to the recognized set.

- Finding 4 (NVRTC sentinel): test_program caught ImportError for
  nvrtc but cuda_utils already imports nvrtc at module level, so the ImportError
  catch was dead code. Import nvrtc directly alongside CUDAError/handle_return and
  keep only the loader/symbol catches around nvrtcVersion().

- Finding 5 (provenance): move the is_gl_context_unavailable tests from
  cuda_python_test_helpers/tests/test_graphics.py to
 cuda_core/tests/test_helpers.py with
 human_reviewed markers, so
  CI actually runs them. Drop the source file.
