+++ cuda_core/tests/test_program.py
@@ -15,7 +15,7 @@ from cuda.core import _linker
 from cuda.core._device import Device
 from cuda.core._module import Kernel, ObjectCode
 from cuda.core._program import Program, ProgramOptions
-from cuda.core._utils.cuda_utils import CUDAError, handle_return
+from cuda.core._utils.cuda_utils import CUDAError, handle_return, nvrtc
 from cuda.core.typing import CompilerBackendType, PCHStatusType
 from cuda.pathfinder import DynamicLibNotFoundError
 
@@ -39,15 +39,6 @@ nvvm_available = pytest.mark.skipif(
     not _is_nvvm_available(), reason="NVVM not available (libNVVM not found or cuda-bindings < 12.9.0)"
 )
 
-# nvrtc is imported lazily (libnvrtc loads on first API call), so an import
-# failure means cuda.bindings is missing; a library/symbol failure surfaces on
-# nvrtcVersion() below. Establish a sentinel so a missing library does not
-# turn into a NameError on the nvrtcVersion() call.
-try:
-    from cuda.core._utils.cuda_utils import nvrtc
-except ImportError:
-    nvrtc = None
-
 
 def _get_nvrtc_version_for_tests():
     """
@@ -57,8 +48,6 @@ def _get_nvrtc_version_for_tests():
         int: Version in format major * 1000 + minor * 100 (e.g., 13200 for CUDA 13.2)
         None: If NVRTC is not available
     """
-    if nvrtc is None:
-        return None
     try:
         nvrtc_major, nvrtc_minor = handle_return(nvrtc.nvrtcVersion())
         return nvrtc_major * 1000 + nvrtc_minor * 100
